+++ src/main.rs
@@ -1,6 +1,9 @@
 #[macro_use]
 extern crate rocket;
 
+#[macro_use]
+extern crate lazy_static;
+
 mod assets;
 mod error;
 mod page;