+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-WIP
-===
+0.4.1 (2022-04-16)
+=====
 
 * Update search index on document change [#41](https://github.com/AustinWise/smeagol/issues/41)
 * Basic cross-site request forgery protection [#43](https://github.com/AustinWise/smeagol/issues/43)